last() function implementation added
Ignore-this: b6e2cbf4b5f748248483747f7bbb789f

darcs-hash:20091019083810-48308-9072d1d0e63e5a83320fe050af2abd1086fc65e5.gz

+++ kernel/tr/executor/xqops/PPReturn.cpp
@@ -152,7 +152,7 @@ void PPReturn::do_next(tuple &t, var_dsc dsc, var_c_id id)
     if (p.svc->at(id))
     {
         p.svc->at(id) = false;
-        t.copy(dsc == pos_dsc ? tuple_cell::atomic((__int64)pos)
+        t.copy(dsc == pos_dsc ? tuple_cell::atomic(pos)
                               : source.cells[p.tuple_pos]);
     }
 	else
